make reposync more verbose and fix bz 1399235
Until now we tried deducing the RPM name in reposync errors
from the line printing 'no more mirrors to try', however in the BZ
1399235, it is actually a combination of the RPM spec file parameters, one
of them being the 'Release' parameter, which is printed to stdout.
This could have been due to a change in yum/yum-utils, or we are
now seeing a different error.

This patch attempts to make the 'reposync' stage more robust and
verbose:

1. New regex lookups were added according to the BZ, the old regex
(probably due to yum/yum-utils changes) is looking for a different
error and also sometimes matched incorrect RPM names(which starting with
'^M' sequence).
It attempts to look RPM names according to the fedora guidelines:
https://fedoraproject.org/wiki/How_to_create_an_RPM_package

2. Defected packages lookup will be done inside python instead
of GNU find and when deleting a file, it will be logged and
also printed to to the user(under the reposync stage). Note that
if no package was was deleted, the output will be the same as
before.

3. reposync will be called for each repoid separately, this will
allow to easily trace which repository caused the error.

bz: https://bugzilla.redhat.com//show_bug.cgi?id=1399235

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>